// src/app/interfaces/pokemon.ts
export interface Ability2 {
    name: string;
    url: string;
  }
  
  export interface Ability {
    is_hidden: boolean;
    slot: number;
    ability: Ability2;
  }
  
  export interface Form {
    name: string;
    url: string;
  }
  
  export interface Version {
    name: string;
    url: string;
  }
  
  export interface GameIndice {
    game_index: number;
    version: Version;
  }
  
  export interface Item {
    name: string;
    url: string;
  }
  
  export interface Version2 {
    name: string;
    url: string;
  }
  
  export interface VersionDetail {
    rarity: number;
    version: Version2;
  }
  
  export interface HeldItem {
    item: Item;
    version_details: VersionDetail[];
  }
  
  export interface Move2 {
    name: string;
    url: string;
  }
  
  export interface VersionGroup {
    name: string;
    url: string;
  }
  
  export interface MoveLearnMethod {
    name: string;
    url: string;
  }
  
  export interface VersionGroupDetail {
    level_learned_at: number;
    version_group: VersionGroup;
    move_learn_method: MoveLearnMethod;
  }
  
  export interface Move {
    move: Move2;
    version_group_details: VersionGroupDetail[];
  }
  
  export interface Species {
    name: string;
    url: string;
  }
  
  export interface DreamWorld {
    front_default: string;
    front_female?: any;
  }
  
  export interface Home {
    front_default: string;
    front_female?: any;
    front_shiny: string;
    front_shiny_female?: any;
  }
  
  export interface OfficialArtwork {
    front_default: string;
  }
  
  export interface Other {
    dream_world: DreamWorld;
    home: Home;
    "official-artwork": OfficialArtwork;
  }
// ...
  export enum Versions {
    "generation-i" = "GenerationI",
    "generation-ii" = "GenerationIi",
    "generation-iii" = "GenerationIii",
    "generation-iv" = "GenerationIv",
    "generation-v" = "GenerationV",
    "generation-vi" = "GenerationVi",
    "generation-vii" = "GenerationVii",
    "generation-viii" = "GenerationViii"
  }
  
  export interface Sprites {
    back_default: string;
    back_female?: any;
    back_shiny: string;
    back_shiny_female?: any;
    front_default: string;
    front_female?: any;
    front_shiny: string;
    front_shiny_female?: any;
    other: Other;
    versions: Versions;
  }
  
  export interface StatDetail {
    name: string;
    url: string;
  }
  
  export interface Stat {
    base_stat: number;
    effort: number;
    stat: StatDetail; 
  }
  
  export interface Type {
    slot: number;
    type: TypeDetail;
  }
  
  export interface TypeDetail {
    name: string;
    url: string;
  }
  
  export interface Generation {
    name: string;
    url: string;
  }
  
  
  export interface PastType {
    generation: Generation;
    types: Type[];
  }
  
  export interface Pokemon {
    id: number;
    name: string;
    base_experience: number;
    height: number;
    is_default: boolean;
    order: number;
    weight: number;
    abilities: Ability[];
    forms: Form[];
    game_indices: GameIndice[];
    held_items: HeldItem[];
    location_area_encounters: string;
    moves: Move[];
    species: Species;
    sprites: Sprites;
    stats: Stat[];
    types: Type[];
    past_types: PastType[];
  }